feat(restapi): merging validation functions
