+++ src/dioptra/restapi/v1/workflows/controller.py
@@ -34,14 +34,11 @@ from .schema import (
     ResourceImportSchema,
     SignatureAnalysisOutputSchema,
     SignatureAnalysisSchema,
-    ValidateEntrypointRequestSchema,
-    ValidateEntrypointResponseSchema,
 )
 from .service import (
     DraftCommitService,
     ResourceImportService,
     SignatureAnalysisService,
-    ValidateEntrypointService,
 )
 
 LOGGER: BoundLogger = structlog.stdlib.get_logger()
@@ -155,47 +152,3 @@ class DraftCommitEndpoint(Resource):
             request_id=str(uuid.uuid4()), resource="DraftCommit", request_type="POST"
         )  # noqa: F841
         return self._draft_commit_service.commit_draft(draft_id=id, log=log)
-
-
-@api.route("/validateEntrypoint")
-class ValidateEntrypointEndpoint(Resource):
-    @inject
-    def __init__(
-        self, validate_entrypoint_service: ValidateEntrypointService, *args, **kwargs
-    ) -> None:
-        """Initialize the workflow resource.
-
-        All arguments are provided via dependency injection.
-
-        Args:
-            entrypoint_validate_service: An EntrypointValidateService object.
-        """
-        self._validate_entrypoint_service = validate_entrypoint_service
-        super().__init__(*args, **kwargs)
-
-    @login_required
-    @accepts(schema=ValidateEntrypointRequestSchema, api=api)
-    @responds(schema=ValidateEntrypointResponseSchema, api=api)
-    def post(self):
-        """Validates the proposed inputs for an entrypoint."""
-        log = LOGGER.new(
-            request_id=str(uuid.uuid4()),
-            resource="ValidateEntrypoint",
-            request_type="POST",
-        )
-        parsed_obj = request.parsed_obj  # pyright: ignore
-        group_id = parsed_obj["group_id"]
-        task_graph = parsed_obj["task_graph"]
-        artifact_graph = parsed_obj.get("artifact_graph", "")
-        plugin_snapshot_ids = parsed_obj["plugin_snapshot_ids"]
-        parameters = parsed_obj["parameters"]
-        artifact_parameters = parsed_obj.get("artifacts", [])
-        return self._validate_entrypoint_service.validate(
-            group_id=group_id,
-            task_graph=task_graph,
-            artifact_graph=artifact_graph,
-            plugin_snapshot_ids=plugin_snapshot_ids,
-            entrypoint_parameters=parameters,
-            entrypoint_artifacts=artifact_parameters,
-            log=log,
-        )
